+++ src/services/index.ts
@@ -1,5 +1,6 @@
 export * as ProjectService from './projects.service'
 export * as SiteService from './sites.api.service'
 export * as MockUpSiteService from './sites.mock.service'
-export * as SpeciesService from './species.service'
+export * as SpeciesService from './species.api.service'
+export * as MockupSpeciesService from './species.mock.service'
 export * as VuexService from './vuex.service'